fix: resolve build failures across multiple CI configurations
- Fix simd_nn_ops.cpp: add immintrin.h include, fix CPUFeatureDetector namespace (cpu:: -> cpu::simd::)
- Fix memory_pool_ops.cpp: add missing <numa.h> include, fix numa_available() check
- Fix algebra_detail.hpp: fix sign-compare warning (Size vs int) with explicit casts
- Fix determinism.cpp: change aggregate init from () to {} for Apple Clang C++20 compat
- Fix hardware_probe.cpp: replace HW_PER_CPU with sysctlbyname(hw.physicalcpu) for macOS compat

+++ src/include/nerve/algebra/detail/algebra_detail.hpp
@@ -105,7 +105,7 @@ public:
         int max_d = -1;
         for (const auto &s : simplices_)
         {
-            if (s.dimension() > max_d)
+            if (static_cast<int>(s.dimension()) > max_d)
                 max_d = s.dimension();
         }
         return max_d;
@@ -115,7 +115,7 @@ public:
         std::vector<Simplex> result;
         for (const auto &s : simplices_)
         {
-            if (s.dimension() == dim)
+            if (static_cast<int>(s.dimension()) == dim)
                 result.push_back(s);
         }
         return result;
